pyproject.toml
[tool.poetry]
name = "grz-service-oex"
version = "0.1.0"
description = "GRZ Service - One Exhibition"
authors = ["Eric Lee <ericstone.dev@gmail.com>"]

[tool.poetry.dependencies]
Django = "^3.2.12"
Pillow = "^8.2.0"
aioredis = "2.0.1"
argon2-cffi = "^21.3.0"
boto3 = "^1.17.93"
celery = "^5.2.5"
cryptography = "^37.0.0"
django-celery-beat = "^2.2.1"
django-celery-results = "^2.3.0"
django-cors-headers = "^3.10.1"
django-environ = "^0.8.1"
django-extensions = "^3.1.5"
django-filter = "^21.1"
django-model-utils = "^4.2.0"
django-redis = "^5.2.0"
django-storages = "^1.12.3"
django-taggit = "^2.1.0"
djangorestframework = "^3.13.1"
fastapi = "^0.75.1"
fastapi-cache2 = { version = "^0.1.3.4", extras = ["redis"] }
hiredis = "^2.0.0"
httpx = "^0.22.0"
openpyxl = "^3.0.7"
pandas = "^1.4.1"
psycopg2-binary = "^2.9.3"
python = "^3.9"
python-jose = { version = "^3.3.0", extras = ["cryptography"] }
python-multipart = "^0.0.5"
python-slugify = "^6.1.2"
redis = "^4.2.1"
requests = "^2.27.1"
shortuuid = "^1.0.8"
supervisor = "^4.2.4"
uvicorn = "^0.17.6"
xlrd = "^2.0.1"
ks-shared = {git = "https://github.com/ericstone57/ks_shared.git", rev = "main"}
djantic = "^0.7.0"
tablestore = "^5.2.1"
django-json-widget = "^1.1.1"
django-simpleui = "^2022.3.19"
django-ckeditor = "^6.3.0"

[tool.poetry.dev-dependencies]
django-debug-toolbar = "^3.2.4"
factory-boy = "^3.2.1"
jupyterlab = "^3.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
